reorder
it seems to make sense to have linux and osx first since a lot of users
use those, and windows at the end because well, it's windows. but it
seemed odd to start with osx, so moved linux up.

alphabetical would maybe be better, but then android comes first, which
just feels weird since it's such a niche port. also the linux distro
sublist has a good reason not to be alphabetical, eg docker and conda
are on there as kind of last resorts